Move CommentCell and CommentCellViewController to their own files
