+++ EssentialApp/EssentialApp/Helpers/CombineHelpers.swift
