+++ EssentialApp/EssentialApp/Helpers/WeakRefVirtualProxy.swift
