+++ EssentialFeed/EssentialFeediOS/Comment UI/Views/CommentCell.swift
@@ -1,0 +1,16 @@
+//
+//  CommentCell.swift
+//  EssentialFeediOS
+//
+//  Created by Khoi Nguyen on 5/1/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+public class CommentCell: UITableViewCell {
+	@IBOutlet public var authorLabel: UILabel?
+	@IBOutlet public var commentLabel: UILabel?
+	@IBOutlet public var timestampLabel: UILabel?
+}
